Refactor #16 #31 : 토큰 예치 이벤트 파일 경로 변경
