+++ src/main/java/com/ddiring/Backend_BlockchainConnector/domain/event/deposit/DepositSucceededEvent.java
@@ -1,4 +1,4 @@
-package com.ddiring.Backend_BlockchainConnector.domain.event.trade;
+package com.ddiring.Backend_BlockchainConnector.domain.event.deposit;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -6,47 +6,44 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 import java.time.Instant;
-import java.time.LocalDateTime;
 
 @Getter
 @Builder
 @NoArgsConstructor(access = lombok.AccessLevel.PROTECTED)
 @AllArgsConstructor
-public class DepositFailedEvent {
+public class DepositSucceededEvent {
     public static final String TOPIC = "DEPOSIT";
 
     private String eventId;
     private String eventType;
     private Instant timestamp;
 
-    private DepositFailedPayload payload;
+    private DepositSucceededPayload payload;
 
     @Getter
     @Builder
     @NoArgsConstructor(access = lombok.AccessLevel.PROTECTED)
     @AllArgsConstructor
-    public static class DepositFailedPayload {
+    public static class DepositSucceededPayload {
         private Long sellId;
         private String status;
         private String sellerAddress;
         private Long sellerTokenAmount;
-        private String errorMessage;
     }
 
-    public static DepositFailedEvent of(Long sellId, String sellerAddress, Long sellerTokenAmount, String errorMessage) {
+    public static DepositSucceededEvent of(Long sellId, String sellerAddress, Long sellerTokenAmount) {
         String uuid = java.util.UUID.randomUUID().toString();
-        String eventType = TOPIC + ".FAILED";
+        String eventType = TOPIC + ".SUCCEEDED";
 
-        return DepositFailedEvent.builder()
+        return DepositSucceededEvent.builder()
                 .eventId(uuid)
                 .eventType(eventType)
                 .timestamp(Instant.now())
-                .payload(DepositFailedPayload.builder()
+                .payload(DepositSucceededPayload.builder()
                         .sellId(sellId)
-                        .status("FAILED")
+                        .status("SUCCEEDED")
                         .sellerAddress(sellerAddress)
                         .sellerTokenAmount(sellerTokenAmount)
-                        .errorMessage(errorMessage)
                         .build()
                 )
                 .build();
